refactor(web): split managed agent configuration modules
