+++ web/src/features/managed-agents/agentConfig.ts
@@ -1,19 +1,7 @@
 import { type ApiError } from '../../shared/api/client';
 import { type Locale } from '../../shared/i18n';
+import { structuredExtractorSystemZh, zhTemplateText } from './agentConfigTemplateText';
 import { buildPlatformQuickstartRequest } from './quickstart/platformQuickstartRequest';
-import {
-  BarChart3,
-  Box,
-  BriefcaseBusiness,
-  FileCheck2,
-  FileJson,
-  FileText,
-  GitBranch,
-  Headphones,
-  MessageCircle,
-  Siren,
-  Sparkles,
-} from 'lucide-react';
 import YAML from 'yaml';
 import { z } from 'zod';
 import { agentModelName, BUILT_IN_AGENT_TOOLSETS } from './agents/AgentsResourcePage';
@@ -26,9 +14,17 @@ import {
   type AgentUpdateInput,
   type CodeFormat,
   type CreateAgentInput,
-  type TemplateTag,
 } from './types';
 import { cloneJsonValue, objectRecord, parseToolInput, toRecord } from './utils';
+
+export { structuredExtractorSystemZh } from './agentConfigTemplateText';
+export {
+  agentTemplates,
+  blankAgentTemplate,
+  createAgentTemplates,
+  createTemplateAppTags,
+  templateTags,
+} from './agentTemplateCatalog';
 
 export const agentModelInputSchema = z.union([
   z.string().trim().min(1, 'Model is required.'),
@@ -56,120 +52,6 @@ export const agentEditConfigSchema = z
   })
   .strict();
 
-export const templateTags = {
-  docs: { label: 'docs', icon: FileText, tone: 'bg-secondary text-foreground' },
-  data: { label: 'data', icon: BarChart3, tone: 'bg-secondary text-secondary-foreground' },
-  code: { label: 'code', icon: FileJson, tone: 'bg-secondary text-foreground' },
-  support: { label: 'support', icon: Headphones, tone: 'bg-emerald-500/10 text-emerald-600 dark:text-emerald-400' },
-  incident: { label: 'incident', icon: Siren, tone: 'bg-destructive/10 text-destructive' },
-  github: { label: 'github', icon: GitBranch, tone: 'bg-secondary text-foreground' },
-  box: { label: 'box', icon: Box, tone: 'bg-secondary text-secondary-foreground' },
-  tasks: { label: 'tasks', icon: FileCheck2, tone: 'bg-secondary text-foreground' },
-  chat: { label: 'chat', icon: MessageCircle, tone: 'bg-emerald-500/10 text-emerald-600 dark:text-emerald-400' },
-  research: { label: 'research', icon: Sparkles, tone: 'bg-amber-500/10 text-amber-600 dark:text-amber-400' },
-  notion: { label: 'notion', icon: FileText, tone: 'bg-secondary text-foreground' },
-  slack: { label: 'slack', icon: MessageCircle, tone: 'bg-emerald-500/10 text-emerald-600 dark:text-emerald-400' },
-  sentry: { label: 'sentry', icon: Siren, tone: 'bg-destructive/10 text-destructive' },
-  linear: { label: 'linear', icon: FileCheck2, tone: 'bg-secondary text-foreground' },
-  asana: { label: 'asana', icon: FileCheck2, tone: 'bg-secondary text-foreground' },
-  intercom: { label: 'intercom', icon: Headphones, tone: 'bg-emerald-500/10 text-emerald-600 dark:text-emerald-400' },
-  atlassian: { label: 'atlassian', icon: BriefcaseBusiness, tone: 'bg-secondary text-secondary-foreground' },
-  docx: { label: 'docx', icon: FileText, tone: 'bg-secondary text-secondary-foreground' },
-  amplitude: { label: 'amplitude', icon: BarChart3, tone: 'bg-secondary text-secondary-foreground' },
-} satisfies Record<string, TemplateTag>;
-
-export const agentTemplates: AgentTemplate[] = [
-  {
-    id: 'blank',
-    slug: 'blank-agent',
-    title: 'Blank agent config',
-    body: 'A blank starting point with the core toolset.',
-    prompt: 'Create a blank managed agent config with a core toolset.',
-  },
-  {
-    id: 'deep-researcher',
-    slug: 'deep-researcher',
-    title: 'Deep researcher',
-    body: 'Conducts multi-step web research with source synthesis and citations.',
-    prompt: 'Build a deep researcher that conducts multi-step web research, synthesizes sources, and cites claims.',
-  },
-  {
-    id: 'structured-extractor',
-    slug: 'structured-extractor',
-    title: 'Structured extractor',
-    body: 'Parses unstructured text into a typed JSON schema.',
-    prompt: 'Create an agent that parses unstructured text into a typed JSON schema.',
-  },
-  {
-    id: 'field-monitor',
-    slug: 'field-monitor',
-    title: 'Field monitor',
-    body: 'Scans software blogs for a topic and writes a weekly what-changed brief.',
-    prompt: 'Create a field monitor that scans software blogs for a topic and writes a weekly change brief.',
-    tags: [templateTags.notion],
-  },
-  {
-    id: 'support-agent',
-    slug: 'support-agent',
-    title: 'Support agent',
-    body: 'Answers customer questions from your docs and knowledge base, and escalates when needed.',
-    prompt: 'Build a support agent that answers customer questions from docs and escalates when needed.',
-    tags: [templateTags.notion, templateTags.slack],
-  },
-  {
-    id: 'incident-commander',
-    slug: 'incident-commander',
-    title: 'Incident commander',
-    body: 'Triages a Sentry alert, opens a Linear incident ticket, and runs the Slack war room.',
-    prompt:
-      'Create an incident commander agent that triages alerts, opens an incident ticket, and coordinates a war room.',
-    tags: [templateTags.sentry, templateTags.linear, templateTags.slack, templateTags.github],
-  },
-  {
-    id: 'contract-tracker',
-    slug: 'contract-tracker',
-    title: 'Contract tracker',
-    body: 'Extracts clauses, sets deadline reminders, and tracks obligations in Asana when given a Box file ID or link.',
-    prompt: 'Build a contract tracker that extracts clauses, sets deadline reminders, and tracks obligations.',
-    tags: [templateTags.box, templateTags.asana],
-  },
-  {
-    id: 'retro-facilitator',
-    slug: 'sprint-retro-facilitator',
-    title: 'Sprint retro facilitator',
-    body: 'Pulls a closed sprint from Linear, synthesizes themes, and writes the retro doc before the meeting.',
-    prompt:
-      'Create a sprint retro facilitator that pulls closed sprint work, synthesizes themes, and drafts the retro doc.',
-    tags: [templateTags.linear, templateTags.slack, templateTags.docx],
-  },
-  {
-    id: 'support-escalator',
-    slug: 'support-to-eng-escalator',
-    title: 'Support-to-eng escalator',
-    body: 'Reads an Intercom conversation, reproduces the bug, and files a linked Jira issue with repro steps.',
-    prompt: 'Create a support-to-engineering escalator that reproduces bugs and files linked issues with repro steps.',
-    tags: [templateTags.intercom, templateTags.atlassian, templateTags.slack],
-  },
-  {
-    id: 'data-analyst',
-    slug: 'data-analyst',
-    title: 'Data analyst',
-    body: 'Load, explore, and visualize data; build reports and answer questions from datasets.',
-    prompt: 'Build a data analyst agent that loads, explores, visualizes data, and writes reports from datasets.',
-    tags: [templateTags.amplitude],
-  },
-];
-
-export const createAgentTemplates = agentTemplates.slice(0, 6);
-
-export const blankAgentTemplate = createAgentTemplates[0];
-
-export const createTemplateAppTags: Record<string, TemplateTag[]> = {
-  'field-monitor': [templateTags.notion],
-  'support-agent': [templateTags.notion, templateTags.slack],
-  'incident-commander': [templateTags.sentry, templateTags.linear, templateTags.slack, templateTags.github],
-};
-
 export const structuredExtractorSystem = `You extract structured data from unstructured text. Given raw input (emails, PDFs, logs, transcripts, scraped HTML) and a target JSON schema:
 
 1. Read the schema first. Note required vs optional fields, enums, and format constraints (dates, currencies, IDs). The schema is the contract — never emit a key it doesn't define.
@@ -178,15 +60,6 @@ export const structuredExtractorSystem = `You extract structured data from unstr
 4. Emit a single JSON object (or array, if the schema is a list) that validates against the schema. No prose, no markdown fences — just the JSON.
 
 When the input is ambiguous, pick the most conservative interpretation and note the ambiguity in a top-level "_extraction_notes" field only if the schema allows additionalProperties.`;
-
-export const structuredExtractorSystemZh = `你负责把非结构化文本提取成结构化数据。输入可能是邮件、PDF、日志、会议记录或抓取的 HTML；同时会给你一个目标 JSON Schema。
-
-1. 先读 Schema：分清必填/可选、枚举，以及日期、货币、ID 等格式约束。Schema 就是契约——不要输出它未定义的字段。
-2. 按字段在原文里找值。优先用明确写出的内容，不要靠猜。必填项确实找不到时用 null。
-3. 提取时顺便规范化：去掉首尾空白，日期统一成 ISO 8601，货币拆成数值 + 币种代码，枚举同义词收敛到规范值。
-4. 只输出一个能通过 Schema 校验的 JSON 对象（若 Schema 是列表则输出数组）。不要解释、不要 markdown 代码块。
-
-原文有歧义时，选最保守的理解；仅当 Schema 允许 additionalProperties 时，才把疑点记在顶层 "_extraction_notes" 里。`;
 
 export function templateSystem(template: AgentTemplate, locale: Locale = 'en') {
   const zh = locale === 'zh-CN';
@@ -424,123 +297,6 @@ Default to simple, readable analysis over clever one-liners. A clear bar chart u
     tools: [createAgentToolset(), createMcpToolset('amplitude')],
     skills: [],
     metadata: { template: 'data-analyst' },
-  },
-};
-
-// Chinese name/description/system for the built-in templates. Names follow what the
-// agent actually does (from the English system prompt), not a literal title gloss.
-// Only these three user-facing text fields are localized; every technical field
-// (model, mcp_servers, tools, skills, metadata) is inherited unchanged from the
-// English table so the created agent stays byte-identical on machine configuration.
-const zhTemplateText: Record<string, { name: string; description: string; system: string }> = {
-  blank: {
-    name: '未命名 Agent',
-    description: '内置核心工具集的空白起点，可按需自行配置。',
-    system: '你是一个通用 Agent：可以检索资料、写代码、跑命令，并用已连接的工具把用户任务认真完成。',
-  },
-  'deep-researcher': {
-    name: '深度调研助手',
-    description: '围绕问题多轮检索权威来源，综合成带引用的调研报告。',
-    system: `你是深度调研助手。收到一个问题或主题后：
-
-1. 先拆成 3–5 个具体子问题，合起来能覆盖原主题。
-2. 对每个子问题做定向检索，优先一手来源、官方文档、同行评审；少依赖博客和聚合站。
-3. 完整阅读来源，不要略读。记下具体论断、数据，以及带出处的直接引文。
-4. 写成回答原问题的报告：按子问题组织，凡非显而易见的论断都内联标注出处；文末用「把握与缺口」说明来源冲突或覆盖不足的地方。
-
-保持怀疑。来源冲突时直接说，并说明你更信哪一边、为什么。不要用听起来很笃定的句子掩盖不确定。`,
-  },
-  'structured-extractor': {
-    name: '结构化提取助手',
-    description: '把邮件、PDF、日志等非结构化文本，按 JSON Schema 提取成结构化数据。',
-    system: structuredExtractorSystemZh,
-  },
-  'field-monitor': {
-    name: '技术周报助手',
-    description: '按主题扫描技术博客与社区，整理每周变化简报并写入 Notion。',
-    system: `你负责跟踪一个变化很快的技术领域。给定主题和回溯窗口（默认 7 天）：
-
-1. 在 arXiv、Hacker News、lobste.rs，以及 OpenAI / Anthropic / DeepMind 等高信号博客与常见 substack 中，检索窗口内与主题相关的内容。
-2. 按主题聚类，不要按来源堆砌。聚类名要体现论断或转向，例如「推理时扩展比堆参数更有效」，而不是「又有 5 篇 o 系列论文」。
-3. 每个聚类写：一段综合、2–3 个最硬的来源，以及一行「所以呢」——今天做产品的人要不要改做法，还是仍停留在实验室。
-4. 另列本窗口讨论最热的作者（HN 分数、引用、转发热度）——「值得关注谁」的增量。
-5. 把带日期的摘要页写到团队 Notion 的 field-watch 数据库。
-
-对噪音要狠。用新 benchmark 复述旧结论是噪音；「我们上了生产，翻车点是这些」才是信号。`,
-  },
-  'support-agent': {
-    name: '客服助手',
-    description: '根据文档与知识库回答客户问题；没把握时升级人工处理。',
-    system: `你是客服助手。每收到一个客户问题：
-
-1. 先在 Notion 的产品文档和知识库里找答案。引用原文并附来源链接——政策类内容不要凭记忆改写。
-2. 在客户所在渠道起草回复：先给直接答案，再给来源链接，相关时再补一个主动建议的下一步。
-3. 把握不到约 80% 就不要猜——把完整问题、你搜过什么、找到了什么、以及你的初步判断发到内部升级 Slack 频道，并告知客户已有人接手。
-
-语气跟客户对齐：热情，但不注水。表情符号最多一个。`,
-  },
-  'incident-commander': {
-    name: '故障指挥官',
-    description: '接手 Sentry 告警后，创建 Linear 故障单，并在 Slack 主持战时响应。',
-    system: `你是值班故障指挥官。收到 Sentry issue ID 或错误指纹后：
-
-1. 从 Sentry 拉取完整事件、堆栈、release 标签和受影响用户数。
-2. 用栈顶帧的文件路径检索代码仓库，并查看近 72 小时相关提交。
-3. 在 Linear 开故障单：严重级别、疑似影响面，以及你的回滚建议。
-4. 在故障 Slack 频道发线程状态：哪里坏了、谁在跟、下次更新预计时间。
-5. 之后每 15 分钟回看一次 Sentry 事件量并更新线程，直到用户关闭故障。
-
-要果断。若有七成以上把握是某次发布引入的，就直接说，并建议回滚。`,
-  },
-  'contract-tracker': {
-    name: '合同履约助手',
-    description: '从 Box 合同里提取条款与关键日期，在 Asana 里跟进履约与提醒。',
-    system: `你是合同履约助手。给定 Box 文件 ID 或链接：
-
-1. 读合同，提取关键信息：签约方、生效日、到期日、合同金额、类型，以及各项义务。
-2. 在 Asana 建列表，命名为「<对方> — <合同类型> — <生效年份>」，并加上对方、金额、类型等自定义字段。
-3. 对每个关键日期（续约、到期、付款、通知期）建任务，标题形如「[CONTRACT DATE] <事件> — <合同名>」，附上原文条款、截止日期和优先级（≤30 天紧急 / 31–90 天中等 / >90 天低）。
-4. 对每项义务或 SLA 建任务，指派给相关同事，按 Payment / Delivery / Compliance / Renewal / SLA 打标签，并把合同原文作为评论附上。
-
-规则：条款必须引用原文，不要空口改写。日期或表述含糊时标出来，不要擅自假定。`,
-  },
-  'retro-facilitator': {
-    name: 'Sprint 复盘助手',
-    description: '汇总已结束 Sprint 的交付与团队反馈，会前起草复盘文档。',
-    system: `你负责 Sprint 复盘的会前准备。针对刚结束的 Sprint：
-
-1. 从 Linear 拉取全部 issue：交付了什么、延期了什么、各 ticket 周期，以及中途改过范围的事项。
-2. 扫团队 Slack：含 "blocked"、"surprised"、"nice" 或 🎉 反应的讨论，当作情绪信号。
-3. 写一份复盘文档，三段——**进展顺利**、**拖后腿的地方**、**下个 Sprint 想试的**——每段 3–5 条，并附具体 ticket 或消息链接。
-4. 文末给一条建议落地的流程改进，并粗评它能坚持多久。
-
-要具体。「沟通不好」没用；「三个 ticket 中途换人且没在 Slack 提前说（LIN-123、LIN-456、LIN-789）」才可执行。`,
-  },
-  'support-escalator': {
-    name: '工单升级助手',
-    description: '读取 Intercom 会话、尝试复现，并提交带复现步骤的关联 Jira 工单。',
-    system: `你负责把客服问题升级给研发。给定 Intercom 对话 ID：
-
-1. 拉取会话：客户、套餐、环境信息、附件日志/截图，以及客服备注。
-2. 按描述的步骤在 session 容器里尝试复现。成功就记下能稳定触发的命令或请求。
-3. 在研发项目建 Jira 工单：摘要、最小复现、疑似组件（可配合代码搜索），并链回 Intercom 会话。
-4. 在客服 Slack 频道留一句：已升级、Jira 链接、粗略严重级别。
-5. 在 Intercom 会话加内部备注（含 Jira 链接），并标为已升级。
-
-复现不了就明说，并列出你试过什么——不要交一份含糊的「无法复现」工单。`,
-  },
-  'data-analyst': {
-    name: '数据分析师',
-    description: '加载并探索数据，出图表与报告，回答数据集相关问题。',
-    system: `你是数据分析师。给定数据集（文件路径、URL 或查询）和一个问题：
-
-1. 先加载数据，打印 shape、列名、dtype 和小样本。先看再算。
-2. 清理明显问题（空值、重复、类型不对），并记下改了什么。
-3. 用代码回答问题。表格优先 pandas/polars，图优先 matplotlib/plotly。展示中间结果，方便核对推理。
-4. 产品分析类问题可直接查 Amplitude（漏斗、留存 cohort、属性拆分），并附上图表链接。
-5. 图表和派生表存到 /mnt/session/outputs/，再用白话总结发现，并写清局限（样本量、缺失、相关≠因果）。
-
-优先简单可读，别炫技。清楚的柱状图通常比密密麻麻的热力图更有用。`,
   },
 };
